==> 8.CNV_2X3Fishers_immunge_genes.R <==
library(Rlabkey)
library(tidyverse, lib="~/re_gecip/immune/phalim/R_packages")
library(curl)
library(rjson)
library(RCurl)
library(dplyr)
library(purrr)
library(ggplot2)

#setting working directory
setwd("/home/phalim/re_gecip/immune/phalim/")

DataPR <- read.csv("/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRare_Fisher.test.csv",sep="\t")
DataPC <- read.csv("/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancer_Fisher.test.csv",sep="\t")

DataPR <- DataPR[order(DataPR$p.value),]
DataPC <- DataPC[order(DataPC$p.value),]

PRBH <- DataPR[,c(1,2)]
PCBH <- DataPC[,c(1,2)]
PRBH$PvsR.BH <- p.adjust(PRBH$p.value, method = "BH")
PCBH$PvsC.BH <- p.adjust(PCBH$p.value, method = "BH")
PRBH$PvsR.BON <- p.adjust(PRBH$p.value, method = "bonferroni")
PCBH$PvsC.BON <- p.adjust(PCBH$p.value, method = "bonferroni")

BHcomp <- full_join(PRBH,PCBH, by="X")
write.table(BHcomp, file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/Pvalueadjustment_PIDvscontrols.csv",sep="\t")

#p.value with 2x3 fisher test
ImmuneGene_CNV_CancerATM <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerATM.csv", sep= "\t")
ImmuneGene_CNV_CancerATM <- ImmuneGene_CNV_CancerATM %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerCFH <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerCFH.csv", sep= "\t")
ImmuneGene_CNV_CancerCFH <- ImmuneGene_CNV_CancerCFH %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerCFHR1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerCFHR1.csv", sep= "\t")
ImmuneGene_CNV_CancerCFHR1 <- ImmuneGene_CNV_CancerCFHR1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerCFHR3 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerCFHR3.csv", sep= "\t")
ImmuneGene_CNV_CancerCFHR3 <- ImmuneGene_CNV_CancerCFHR3 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerCFHR4 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerCFHR4.csv", sep= "\t")
ImmuneGene_CNV_CancerCFHR4 <- ImmuneGene_CNV_CancerCFHR4 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerDNASE1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerDNASE1.csv", sep= "\t")
ImmuneGene_CNV_CancerDNASE1 <- ImmuneGene_CNV_CancerDNASE1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerFANCA <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerFANCA.csv", sep= "\t")
ImmuneGene_CNV_CancerFANCA <- ImmuneGene_CNV_CancerFANCA %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerGFI1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerGFI1.csv", sep= "\t")
ImmuneGene_CNV_CancerGFI1 <- ImmuneGene_CNV_CancerGFI1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerIGHG2 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerIGHG2.csv", sep= "\t")
ImmuneGene_CNV_CancerIGHG2 <- ImmuneGene_CNV_CancerIGHG2 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerIL12RB2 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerIL12RB2.csv", sep= "\t")
ImmuneGene_CNV_CancerIL12RB2 <- ImmuneGene_CNV_CancerIL12RB2 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerNFKB1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerNFKB1.csv", sep= "\t")
ImmuneGene_CNV_CancerNFKB1 <- ImmuneGene_CNV_CancerNFKB1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerPLEKHM1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerPLEKHM1.csv", sep= "\t")
ImmuneGene_CNV_CancerPLEKHM1 <- ImmuneGene_CNV_CancerPLEKHM1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerPTPRC <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerPTPRC.csv", sep= "\t")
ImmuneGene_CNV_CancerPTPRC <- ImmuneGene_CNV_CancerPTPRC %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerRAB27A <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerRAB27A.csv", sep= "\t")
ImmuneGene_CNV_CancerRAB27A <- ImmuneGene_CNV_CancerRAB27A %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerTET2 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerTET2.csv", sep= "\t")
ImmuneGene_CNV_CancerTET2 <- ImmuneGene_CNV_CancerTET2 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerUSP18 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerUSP18.csv", sep= "\t")
ImmuneGene_CNV_CancerUSP18 <- ImmuneGene_CNV_CancerUSP18 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerVPS13B <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerVPS13B.csv", sep= "\t")
ImmuneGene_CNV_CancerVPS13B <- ImmuneGene_CNV_CancerVPS13B %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_CancerWDR1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsCancer/Count_lossCNV_PIDvsCancerWDR1.csv", sep= "\t")
ImmuneGene_CNV_CancerWDR1 <- ImmuneGene_CNV_CancerWDR1 %>% remove_rownames %>% column_to_rownames(var="X")

ImmuneGene_CNV_RareATM <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareWDR1.csv", sep= "\t")
ImmuneGene_CNV_RareATM <- ImmuneGene_CNV_RareATM %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareCFH <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareCFH.csv", sep= "\t")
ImmuneGene_CNV_RareCFH <- ImmuneGene_CNV_RareCFH %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareCFHR1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareCFHR1.csv", sep= "\t")
ImmuneGene_CNV_RareCFHR1 <- ImmuneGene_CNV_RareCFHR1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareCFHR3 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareCFHR3.csv", sep= "\t")
ImmuneGene_CNV_RareCFHR3 <- ImmuneGene_CNV_RareCFHR3 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareCFHR4 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareCFHR4.csv", sep= "\t")
ImmuneGene_CNV_RareCFHR4 <- ImmuneGene_CNV_RareCFHR4 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareDNASE1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareDNASE1.csv", sep= "\t")
ImmuneGene_CNV_RareDNASE1 <- ImmuneGene_CNV_RareDNASE1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareFANCA <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareFANCA.csv", sep= "\t")
ImmuneGene_CNV_RareFANCA <- ImmuneGene_CNV_RareFANCA %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareGFI1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareGFI1.csv", sep= "\t")
ImmuneGene_CNV_RareGFI1 <- ImmuneGene_CNV_RareGFI1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareIGHG2 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareIGHG2.csv", sep= "\t")
ImmuneGene_CNV_RareIGHG2 <- ImmuneGene_CNV_RareIGHG2 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareIL12RB2 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareIL12RB2.csv", sep= "\t")
ImmuneGene_CNV_RareIL12RB2 <- ImmuneGene_CNV_RareIL12RB2 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareNFKB1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareNFKB1.csv", sep= "\t")
ImmuneGene_CNV_RareNFKB1 <- ImmuneGene_CNV_RareNFKB1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RarePLEKHM1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRarePLEKHM1.csv", sep= "\t")
ImmuneGene_CNV_RarePLEKHM1 <- ImmuneGene_CNV_RarePLEKHM1 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RarePTPRC <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRarePTPRC.csv", sep= "\t")
ImmuneGene_CNV_RarePTPRC <- ImmuneGene_CNV_RarePTPRC %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareRAB27A <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareRAB27A.csv", sep= "\t")
ImmuneGene_CNV_RareRAB27A <- ImmuneGene_CNV_RareRAB27A %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareTET2 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareTET2.csv", sep= "\t")
ImmuneGene_CNV_RareTET2 <- ImmuneGene_CNV_RareTET2 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareUSP18 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareUSP18.csv", sep= "\t")
ImmuneGene_CNV_RareUSP18 <- ImmuneGene_CNV_RareUSP18 %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareVPS13B <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareVPS13B.csv", sep= "\t")
ImmuneGene_CNV_RareVPS13B <- ImmuneGene_CNV_RareVPS13B %>% remove_rownames %>% column_to_rownames(var="X")
ImmuneGene_CNV_RareWDR1 <- read.csv(file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsRare/Count_lossCNV_PIDvsRareWDR1.csv", sep= "\t")
ImmuneGene_CNV_RareWDR1 <- ImmuneGene_CNV_RareWDR1 %>% remove_rownames %>% column_to_rownames(var="X")

PCR_ATM <- as.matrix(full_join(ImmuneGene_CNV_RareATM,ImmuneGene_CNV_CancerATM))
PCR_CFH <- as.matrix(full_join(ImmuneGene_CNV_RareCFH,ImmuneGene_CNV_CancerCFH))
PCR_CFHR1 <- as.matrix(full_join(ImmuneGene_CNV_RareCFHR1,ImmuneGene_CNV_CancerCFHR1))
PCR_CFHR3 <- as.matrix(full_join(ImmuneGene_CNV_RareCFHR3,ImmuneGene_CNV_CancerCFHR3))
PCR_CFHR4 <- as.matrix(full_join(ImmuneGene_CNV_RareCFHR4,ImmuneGene_CNV_CancerCFHR4))
PCR_DNASE1 <- as.matrix(full_join(ImmuneGene_CNV_RareDNASE1,ImmuneGene_CNV_CancerDNASE1))
PCR_FANCA <- as.matrix(full_join(ImmuneGene_CNV_RareFANCA,ImmuneGene_CNV_CancerFANCA))
PCR_GFI1 <- as.matrix(full_join(ImmuneGene_CNV_RareGFI1,ImmuneGene_CNV_CancerGFI1))
PCR_IGHG2 <- as.matrix(full_join(ImmuneGene_CNV_RareIGHG2,ImmuneGene_CNV_CancerIGHG2))
PCR_IL12RB2 <- as.matrix(full_join(ImmuneGene_CNV_RareIL12RB2,ImmuneGene_CNV_CancerIL12RB2))
PCR_NFKB1 <- as.matrix(full_join(ImmuneGene_CNV_RareNFKB1,ImmuneGene_CNV_CancerNFKB1))
PCR_PLEKHM1 <- as.matrix(full_join(ImmuneGene_CNV_RarePLEKHM1,ImmuneGene_CNV_CancerPLEKHM1))
PCR_PTPRC <- as.matrix(full_join(ImmuneGene_CNV_RarePTPRC,ImmuneGene_CNV_CancerPTPRC))
PCR_RAB27A <- as.matrix(full_join(ImmuneGene_CNV_RareRAB27A,ImmuneGene_CNV_CancerRAB27A))
PCR_TET2 <- as.matrix(full_join(ImmuneGene_CNV_RareTET2,ImmuneGene_CNV_CancerTET2))
PCR_USP18 <- as.matrix(full_join(ImmuneGene_CNV_RareUSP18,ImmuneGene_CNV_CancerUSP18))
PCR_VPS13B <- as.matrix(full_join(ImmuneGene_CNV_RareVPS13B,ImmuneGene_CNV_CancerVPS13B))
PCR_WDR1 <- as.matrix(full_join(ImmuneGene_CNV_RareWDR1,ImmuneGene_CNV_CancerWDR1))

PCR_list <- list(PCR_ATM,PCR_CFH,PCR_CFHR1,PCR_CFHR3,PCR_CFHR4,PCR_DNASE1,PCR_FANCA,
                 PCR_GFI1,PCR_IGHG2,PCR_IL12RB2,PCR_NFKB1,PCR_PLEKHM1,PCR_PTPRC,
                 PCR_RAB27A,PCR_TET2,PCR_USP18,PCR_VPS13B,PCR_WDR1)

Fischer_ATM <- fisher.test(PCR_ATM, simulate.p.value = TRUE)
Fischer_CFH <- fisher.test(PCR_CFH, simulate.p.value = TRUE)
Fischer_CFHR1 <- fisher.test(PCR_CFHR1, simulate.p.value = TRUE)
Fischer_CFHR3 <- fisher.test(PCR_CFHR3, simulate.p.value = TRUE)
Fischer_CFHR4 <- fisher.test(PCR_CFHR4, simulate.p.value = TRUE)
Fischer_DNASE1 <- fisher.test(PCR_DNASE1, simulate.p.value = TRUE)
Fischer_FANCA <- fisher.test(PCR_FANCA, simulate.p.value = TRUE)
Fischer_GFI1 <- fisher.test(PCR_GFI1, simulate.p.value = TRUE)
Fischer_IGHG2 <- fisher.test(PCR_IGHG2, simulate.p.value = TRUE)
Fischer_IL12RB2 <- fisher.test(PCR_IL12RB2, simulate.p.value = TRUE)
Fischer_NFKB1 <- fisher.test(PCR_NFKB1, simulate.p.value = TRUE)
Fischer_PLEKHM1 <- fisher.test(PCR_PLEKHM1, simulate.p.value = TRUE)
Fischer_PTPRC <- fisher.test(PCR_PTPRC, simulate.p.value = TRUE)
Fischer_RAB27A <- fisher.test(PCR_RAB27A, simulate.p.value = TRUE)
Fischer_TET2 <- fisher.test(PCR_TET2, simulate.p.value = TRUE)
Fischer_USP18 <- fisher.test(PCR_USP18, simulate.p.value = TRUE)
Fischer_VPS13B <- fisher.test(PCR_VPS13B, simulate.p.value = TRUE)
Fischer_WDR1 <- fisher.test(PCR_WDR1, simulate.p.value = TRUE)

Fischer_PIDvsControl <- list(Fischer_ATM,Fischer_CFH,Fischer_CFHR1,Fischer_CFHR3,Fischer_CFHR4,Fischer_DNASE1,
                             Fischer_FANCA,Fischer_GFI1,Fischer_IGHG2,Fischer_IL12RB2,Fischer_NFKB1,Fischer_PLEKHM1,
                             Fischer_PTPRC,Fischer_RAB27A,Fischer_TET2,Fischer_USP18,Fischer_VPS13B,Fischer_WDR1)

CNV_PIDvsControlFisher <- data.frame(matrix(unlist(Fischer_PIDvsControl), nrow=length(Fischer_PIDvsControl), byrow=T))

names(CNV_PIDvsControlFisher)[1:4] <- c("p.value", "alternative", "method", "data name")
#rownames(CNV_PIDvsControlFisher)[1:18]<- c("ATM","CFH","CFHR1","CFHR3","CFHR4","DNASE1","FANCA","GFI1","IGHG2","IL12RB2",
#                                                     "NFKB1","PLEKHM1","PTPRC","RAB27A","TET2","USP18","VPS13B","WDR1")
CNV_PIDvsControlFisher <- CNV_PIDvsControlFisher[,c(4,1)]
CNV_PIDvsControlFisher <- CNV_PIDvsControlFisher[order(CNV_PIDvsControlFisher$p.value),]
write.table(CNV_PIDvsControlFisher, file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsControl_Fisher.test.csv", sep= "\t")

Fisher.BH <- read.csv (file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsControl_Fisher.test.csv", sep= "\t")
Fisher.BH$BH.Pvscontrol <- p.adjust(Fisher.BH$p.value , method = "BH", n=18)
write.table(Fisher.BH, file="/home/phalim/re_gecip/immune/phalim/SVCNV/CNV_frequency_statistics/PIDvsControl_Fisher.test.csv", sep= "\t")
